double  to int fix

+++ lib/src/map_location_picker.dart
@@ -13,7 +13,7 @@ import 'autocomplete_view.dart';
 import 'logger.dart';
 
 class MapLocationPicker extends StatefulWidget {
-  final List<double>? radiusList;
+  final List<int>? radiusList;
   final bool isPolygon;
   final bool isPolyline;
   final bool isCircle;
@@ -361,7 +361,8 @@ class _MapLocationPickerState extends State<MapLocationPicker> {
               (e) => Circle(
                   circleId: CircleId(e.key),
                   center: e.value,
-                  radius: widget.radiusList?[int.parse(e.key)] ?? 1.0,
+                  radius:
+                      widget.radiusList?[int.parse(e.key)].toDouble() ?? 1.0,
                   fillColor: Colors.grey),
             )
             .toList() ??
